% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_custom_year.R
\name{convert_date_to_custom_year}
\alias{convert_date_to_custom_year}
\title{Convert date to a year that does not start with the first of january}
\usage{
convert_date_to_custom_year(date, start_month = 3,
  start_day_in_month = 1)
}
\arguments{
\item{date}{the date as an unambiguous character or POSIXct vector}

\item{start_month}{the month in which the current year starts}

\item{start_day_in_month}{the first day in the month where the current year starts}
}
\value{
numeric vector containing the new year value
}
\description{
Convert date to a year that does not start with the first of january
}
\examples{
dates <- c("2018-02-28", "2018-03-01", "2018-03-02")
convert_date_to_custom_year(dates, start_month = 3, start_day = 1)
}
