% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbopvr.R
\docType{package}
\name{inbopvr}
\alias{inbopvr}
\title{\code{inbopvr} package}
\description{
Pakket die enkele functies voor verschillende projecten combineert. Voorlopig gaat dit over de PCI index voor everzwijnen en enkele populatiemodelleringsfuncties.
}
