% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_for_conclict_index_PCI.R
\name{qd_moments}
\alias{qd_moments}
\title{Calculate some base stats on the pci calculation}
\usage{
qd_moments(x, nsim = 1000, func = "qd_pci2", ...)
}
\arguments{
\item{x}{vector with scores of the respondents}

\item{nsim}{amount of simulations}

\item{func}{function to use (as character) "qd_pci1" or "qd_pci2"}

\item{...}{extra arguments (like m and p for qd_pci2)}
}
\value{
Vector of statistics
}
\description{
Simulates pci values as a bootstrap based upon an observed sample
}
\examples{
{
scores <- c(rep(-2,10), rep(-1,10), rep(0,10), rep(1,10), rep(2,10))
qd_moments(scores, nsim = 10000, m = 2, p = 1)
}
}
