% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_for_conclict_index_PCI.R
\name{qd_pci1}
\alias{qd_pci1}
\title{Potential conflict index (first variant)}
\usage{
qd_pci1(x, scale_values = c(-2:2), x_is_table = FALSE)
}
\arguments{
\item{x}{vector with scores of the respondents}

\item{scale_values}{vector with levels; default: -2:2}

\item{x_is_table}{if TRUE, x is table with the distribution of the scores}
}
\value{
PCI-score (potential for conflict index)
}
\description{
questionnaire data analysis: potential conflict index
}
\examples{
{
set.seed(201)
Xv <- sample(-2:2, size = 100, replace = TRUE) #random responses
Yv <- rep(c(-2,2),50) #most extreme difference
Zv <- rep(2,100) #minimal difference
#qd_pci1
qd_pci1(Xv, scale_values = -2:2, x_is_table = FALSE) # 0.4
qd_pci1(Yv, scale_values = -2:2, x_is_table = FALSE) # 1
qd_pci1(Zv, scale_values = -2:2, x_is_table = FALSE) # 0
}
}
