% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_for_conclict_index_PCI.R
\name{qd_pci_2sample_test}
\alias{qd_pci_2sample_test}
\title{Compare 2 pci values}
\usage{
qd_pci_2sample_test(pci1 = NULL, pci2 = NULL, sd1 = NULL, sd2 = NULL)
}
\arguments{
\item{pci1}{precalculated pci first group}

\item{pci2}{precalculated pci second group}

\item{sd1}{precacluated sd for first group on pci}

\item{sd2}{precalculated sd for second group on pci}
}
\value{
Vector of statistics
}
\description{
Performs a test comparable with the unpaired t-test for 2 groups
}
\details{
I am not convinced this is the right way, but that's how it is in the excel application
}
\examples{
{
qd_pci_2sample_test(pci1 = 0.75, pci2 = 0.50, sd1 = 0.06, sd2 = 0.05)
}
}
