% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mark_lifehistory.R
\name{sim_individual}
\alias{sim_individual}
\title{Simulate a lifehistory for an individual with capture, recapture, resighting and ring recovery (close to Barker model)}
\usage{
sim_individual(ringyear, endyear, ring_age = 0, age_based = TRUE, S,
  LL, LD, R, P)
}
\arguments{
\item{ringyear}{the year in the survey on which the animal is ringed}

\item{endyear}{the last year of the survey, assuming it starts by time 1 and there is a recapture every year}

\item{ring_age}{the age on which the animal is ringed}

\item{age_based}{ig TRUE then the parameters will vary by age, otherwise they wil vary by time}

\item{S}{constant or vector containing the survival by year or age depending on paramter age_based}

\item{LL}{constant or vector containing the chance for a sighting between capture evens if the animal survives to the next year}

\item{LD}{constant or vector containing the chance for a sighting between capture evens if the animal survives to the next year}

\item{R}{constant or vector containing the chance for a ring recovery if the animals dies between t and t+1}

\item{P}{constant or vector containing the chance for a recapture at t+1}
}
\value{
a lifehistory string for the aimal
}
\description{
The model assumes that there is a recapture event each fixed period t, t+1, t+2 And that there are alive ring resightings between the timesteps, and/or dead recovery between the time steps
}
\examples{
endyear <- 10
indivs <- expand.grid(Time = 1:1, Id = 1:50, Age = 0:0)
for (i in 1:nrow(indivs)) {
 indivs$ch[i] <-
   sim_individual(ringyear = indivs$Time[i], endyear = endyear, ring_age = indivs$Age[i], age_based = TRUE,
                  S = 0.7, LL = 0.0, LD = 0.0, R = 0.0, P = 0.8)
}
}
